+++ crates/nostr-relay-pool/src/stream.rs
@@ -2,39 +2,21 @@
 // Copyright (c) 2023-2025 Rust Nostr Developers
 // Distributed under the MIT software license
 
-//! Streams
-
 use std::pin::Pin;
 use std::task::{Context, Poll};
 
 use async_utility::futures_util::Stream;
 use tokio::sync::mpsc::Receiver;
 
-/// A wrapper around [`Receiver`] that implements [`Stream`].
 #[derive(Debug)]
-pub struct ReceiverStream<T> {
+pub(crate) struct ReceiverStream<T> {
     inner: Receiver<T>,
 }
 
 impl<T> ReceiverStream<T> {
-    /// Create a new `ReceiverStream`.
     #[inline]
-    pub fn new(recv: Receiver<T>) -> Self {
+    pub(crate) fn new(recv: Receiver<T>) -> Self {
         Self { inner: recv }
-    }
-
-    /// Get back the inner [`Receiver`].
-    #[inline]
-    pub fn into_inner(self) -> Receiver<T> {
-        self.inner
-    }
-
-    /// Closes the receiving half of a channel without dropping it.
-    ///
-    /// Check [`Receiver::close`] to learn more.
-    #[inline]
-    pub fn close(&mut self) {
-        self.inner.close();
     }
 }
 
@@ -43,17 +25,5 @@ impl<T> Stream for ReceiverStream<T> {
 
     fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
         self.inner.poll_recv(cx)
-    }
-}
-
-impl<T> AsRef<Receiver<T>> for ReceiverStream<T> {
-    fn as_ref(&self) -> &Receiver<T> {
-        &self.inner
-    }
-}
-
-impl<T> AsMut<Receiver<T>> for ReceiverStream<T> {
-    fn as_mut(&mut self) -> &mut Receiver<T> {
-        &mut self.inner
     }
 }