+++ crates/nostr-sdk/CHANGELOG.md
@@ -28,6 +28,7 @@
 ### Breaking changes
 
 - Remove `lmdb`, `ndb` and `indexeddb` features (https://github.com/rust-nostr/nostr/pull/1083)
+- Replace `ReceiverStream` with `BoxStream` (https://github.com/rust-nostr/nostr/pull/1087)
 
 ### Added